omg the frames aren't showing
I went display -> current frame, and it showed on the screen. I tried to
playblast but the avi kept getting corrupted or I just couldn't play it,
I did try to playblast the 56 seconds all at once. So instead I had to
save 1368 frames of images but the frame count didn't show up in those??
Why!!!